pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "nnusf"
version = "0.1.0"
description = ""
authors = [
  "Alessandro Candido <candido.ale@gmail.com>",
  "Giacomo Magni <giac.magni@gmail.com>",
  "Tanjona R. Rabemananjara <tanjona.lspc@gmail.com>",
  "Roy Stegeman <roystegeman@live.nl>",
]

[tool.poetry.dependencies]
python = "^3.9,<3.11"
h5py = "^3.7.0"
PyYAML = "^6.0"
yadism = "^0.11.4"
pineappl = "^0.5.0"
matplotlib = "^3.5.2"
seaborn = "^0.11.2"
pandas = "^1.4.2"
rich = "^12.5.1"
scipy = "^1.8.1"
tensorflow = "<2.9.10"
pylint = "^2.14.4"
appdirs = "^1.4.4"
banana-hep = "^0.6.6"
click = "^8.1.3"
greenlet = "^1.1.2"
pendulum = "^2.1.2"
pytzdata = "^2020.1"
SQLAlchemy = "^1.4.39"
termcolor = "1.1.0"
pygit2 = "^1.10.1"
setuptools = "^65.5.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
pytest-cov = "^2.12.1"
ipython = "^8.4.0"
pdbpp = "^0.10.3"

[tool.poetry.scripts]
nnu = "nnusf.cli:command"

[tool.poe.tasks]
coverage = "$BROWSER htmlcov/index.html"
test = "pytest tests"
test-cov = ["test", "coverage"]
lint = "pylint src/**/*.py -E"
lint-warnings = "pylint src/**/*.py --exit-zero"

[tool.pytest.ini_options]
testpaths = ['tests/']
python_files = ['test_*.py', 'benchmark_*.py', 'regression.py']
python_classes = ['Test*']
python_functions = ['test_*']
addopts = [
  '--cov=nnusf',
  '--cov-report=html',
  '--cov-report=xml',
  '--strict-markers',
]
